Added improvements in shop
