+++ app/YetiForce/Shop/Product/YetiForcePremiumSupport.php
@@ -1,0 +1,40 @@
+<?php
+/**
+ * YetiForce shop PremiumSupport file.
+ *
+ * @package   App
+ *
+ * @copyright YetiForce Sp. z o.o
+ * @license   YetiForce Public License 3.0 (licenses/LicenseEN.txt or yetiforce.com)
+ * @author    Mariusz Krzaczkowski <m.krzaczkowski@yetiforce.com>
+ */
+
+namespace App\YetiForce\Shop\Product;
+
+/**
+ * YetiForce shop PremiumSupport class.
+ */
+class YetiForcePremiumSupport extends \App\YetiForce\Shop\AbstractBaseProduct
+{
+	/**
+	 * {@inheritdoc}
+	 */
+	public $prices = [
+		'Micro' => 25,
+		'Small' => 50,
+		'Medium' => 100,
+		'Large' => 250,
+	];
+	/**
+	 * {@inheritdoc}
+	 */
+	public $featured = true;
+
+	/**
+	 * {@inheritdoc}
+	 */
+	public function verify($cache = true): bool
+	{
+		return true;
+	}
+}
