+++ languages/en-US/Settings/YetiForce.json
@@ -69,8 +69,6 @@
     "LBL_SHOP_PROCEED_TO_CHECKOUT": "Proceed to payment",
     "LBL_SHOP_NO_COMPANIES_ALERT": "There is no data of the target user (the company that owns the CRM) in the system. The data is required to issue an invoice, it is not possible to make a payment without this data.",
     "LBL_SHOP_NO_COMPANIES_BUTTON": "Fill in the data",
-    "LBL_SHOP_INSTALLINCLOUD": "Install in cloud",
-    "LBL_SHOP_INSTALLINHOSTING": "Install in shared hosting",
     "LBL_SHOP_COMPANY_DATA": "Company data",
     "LBL_SHOP_COMPANY_SIZE_MICRO": "Micro (up to 10)",
     "LBL_SHOP_COMPANY_SIZE_SMALL": "Small (up to 50)",
@@ -78,7 +76,7 @@
     "LBL_SHOP_COMPANY_SIZE_LARGE": "Large (above 250)",
     "LBL_SHOP_DOMAIN_PREFIX": "Domain prefix",
     "LBL_SHOP_INSTALLINCLOUD_DESC": "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Qui tempora nemo repellat tenetur libero nihil enim porro nesciunt laboriosam vel. Repudiandae, sit reprehenderit deserunt hic eius porro dolores vitae numquam!",
-    "LBL_SHOP_INSTALLINHOSTING_DESC": "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Qui tempora nemo repellat tenetur libero nihil enim porro nesciunt laboriosam vel. Repudiandae, sit reprehenderit deserunt hic eius porro dolores vitae numquam!",
+
     "LBL_VULNERABILITIES_OK": "No vulnerabilities found in libraries",
     "LBL_VULNERABILITIES_OK_DESC": "Do not forget to check the vulnerabilities in your libraries periodically.",
     "LBL_VULNERABILITIES_WARNING": "Detected vulnerabilities in libraries",
@@ -90,24 +88,27 @@
     "LBL_SHOP_PREMIUMMODULES": "Premium modules",
     "LBL_SHOP_PREMIUMMODULES_INTRO": "Enable new system capabilities",
     "LBL_SHOP_PREMIUMMODULES_DESC": "Enable new system capabilities",
-    "LBL_SHOP_DONATIONS": "Donate / Support our project",
-    "LBL_SHOP_DONATIONS_INTRO": "Would you like to influence the future of YetiForce? Support its development!",
-    "LBL_SHOP_DONATIONS_DESC": "Would you like to influence the future of YetiForce? Support its development!",
-    "LBL_SHOP_PREMIUMSUPPORT": "Premium support",
-    "LBL_SHOP_PREMIUMSUPPORT_INTRO": "If you do not know how the system works, or you require individual and quick help, this package is for you",
-    "LBL_SHOP_PREMIUMSUPPORT_DESC": "If you do not know how the system works, or you require individual and quick help, this package is for you",
-    "LBL_SHOP_DISABLEBRANDING": "Disable branding",
-    "LBL_SHOP_DISABLEBRANDING_INTRO": "You can now disable the YetiForce footer for a fee.",
-    "LBL_SHOP_DISABLEBRANDING_DESC": "You can now disable the YetiForce footer for a fee.",
+    "LBL_SHOP_YETIFORCEDONATIONS": "Donate / Support our project",
+    "LBL_SHOP_YETIFORCEDONATIONS_INTRO": "Would you like to influence the future of YetiForce? Support its development!",
+    "LBL_SHOP_YETIFORCEDONATIONS_DESC": "Would you like to influence the future of YetiForce? Support its development!",
+    "LBL_SHOP_YETIFORCEPREMIUMSUPPORT": "Premium support",
+    "LBL_SHOP_YETIFORCEPREMIUMSUPPORT_INTRO": "If you do not know how the system works, or you require individual and quick help, this package is for you",
+    "LBL_SHOP_YETIFORCEPREMIUMSUPPORT_DESC": "If you do not know how the system works, or you require individual and quick help, this package is for you",
+    "LBL_SHOP_YETIFORCEDISABLEBRANDING": "Disable branding",
+    "LBL_SHOP_YETIFORCEDISABLEBRANDING_INTRO": "You can now disable the YetiForce footer for a fee.",
+    "LBL_SHOP_YETIFORCEDISABLEBRANDING_DESC": "You can now disable the YetiForce footer for a fee.",
     "LBL_SHOP_INTEGRATOR": "Integrator",
     "LBL_SHOP_INTEGRATOR_INTRO": "Offer the YetiForce system to your clients and gain exclusive perks available only to YetiForce Integrator partners.",
     "LBL_SHOP_INTEGRATOR_DESC": "REQUIREMENTS \n\n 1. Correct registration of all YetiForce installations used by the Partner and the Partner’s clients. \n 2. Owning a GitHub account and active participation in the YetiForce repository. \n\n BENEFITS \n\n 1. Direct support (chat, e-mail, GitHub) from the YetiForce team in terms of advanced problems related to system operation, API and configuration. Support can be provided in Polish or English. Support may also be provided to the Partner’s customers in cases when the Partner will not be able to help its customers. \n 2. The possibility of hiring YetiForce’s employees for large implementations (service available for an additional charge, provided remotely, communication in Polish and English). \n 3. The opportunity to cooperate in tenders (shared customer testimonials, a common list of specialists, support in price estimates). \n 4. Ability to use the “YetiForce Integrator” logo on the Internet. \n 5. Promotion in the partnership catalog on the official website and in the YetiForce system. \n 6. 10% discount on all products offered in the marketplace by the YetiForce producer and other Partners (also for Partner’s own customers). The possibility of booking a project - the discount might be 20% but it will be granted to a Partner which books the project as first. \n 7. The impact on the direction of system development and new functionalities. \n\n  COST \n\n 1. The monthly participation fee for the YetiForce Integrator partnership is 250 €.",
-    "LBL_SHOP_INSTALLINCLOUD": "Installation in the cloud",
-    "LBL_SHOP_INSTALLINCLOUD_INTRO": "",
-    "LBL_SHOP_INSTALLINCLOUD__DESC": "",
-    "LBL_SHOP_INSTALLINHOSTING": "Installation on hosting",
-    "LBL_SHOP_INSTALLINHOSTING_INTRO": "",
-    "LBL_SHOP_INSTALLINHOSTING__DESC": "",
+    "LBL_SHOP_YETIFORCEDEVELOPMENTSUPPORT": "Development support",
+    "LBL_SHOP_YETIFORCEDEVELOPMENTSUPPORT_INTRO": "",
+    "LBL_SHOP_YETIFORCEDEVELOPMENTSUPPORT_DESC": "",
+    "LBL_SHOP_YETIFORCEINSTALLINCLOUD": "Installation in the cloud",
+    "LBL_SHOP_YETIFORCEINSTALLINCLOUD_INTRO": "",
+    "LBL_SHOP_YETIFORCEINSTALLINCLOUD_DESC": "",
+    "LBL_SHOP_YETIFORCEINSTALLINHOSTING": "Installation on hosting",
+    "LBL_SHOP_YETIFORCEINSTALLINHOSTING_INTRO": "",
+    "LBL_SHOP_YETIFORCEINSTALLINHOSTING_DESC": "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Qui tempora nemo repellat tenetur libero nihil enim porro nesciunt laboriosam vel. Repudiandae, sit reprehenderit deserunt hic eius porro dolores vitae numquam!",
     "LBL_SHOP_INFORMATION": "The prices of the products and services below vary depending on the size of the company. The size of the company is determined by the number of system users. The application will automatically adjust the prices to the size of the company each month, according to the following criteria: Micro company (up to 10 users), Small company (up to 50 users), Medium company (up to 250 users), Large company (over 250 users).",
     "LBL_SHOP_RESELLER": "Reseller",
     "LBL_SHOP_RESELLER_INTRO": "Sell YetiForce’s products and services, earn money on promoting YetiForce in your country.",
