+++ languages/en-US/Settings/_Base.json
@@ -113,8 +113,6 @@
     "LBL_PARTNER_ZONE": "Partner zone",
     "LBL_SYSTEM_MONITORING": "System monitoring",
     "LBL_SHOP_MARKETPLACE": "Marketplace",
-    "LBL_SHOP_INSTALLINCLOUD_INTRO": "",
-    "LBL_SHOP_INSTALLINHOSTING_INTRO": "",
     "LBL_SHOP_RENEW": "Renew",
     "LBL_SETTINGS_SHORT_CUT": "My shortcuts",
     "LBL_CALENDAR_LABELS_COLORS": "Colors and calendar",
